Folder renames for webviewPanel
