+++ src/client/common/installer/serviceRegistry.ts
@@ -4,7 +4,7 @@
 
 import { IServiceManager } from '../../ioc/types';
 import { IWebviewPanelProvider } from '../application/types';
-import { WebviewPanelProvider } from '../application/webPanels/webviewPanelProvider';
+import { WebviewPanelProvider } from '../application/webviewPanels/webviewPanelProvider';
 import { ProductType } from '../types';
 import { InstallationChannelManager } from './channelManager';
 import { CondaInstaller } from './condaInstaller';
